+++ packages/apollo-angular/src/types.ts
@@ -1,3 +1,6 @@
 // XXX: 2.0.0-rc.3
 // import {ApolloClientOptions} from 'apollo-client';
 export type ApolloOptions = Record<string, any>;
+export type TypedVariables<T> = {
+  variables?: T;
+};